Move source files under lib and refactor VM commands
