+++ bindings/nodejs/modules/open-intent/lib/api/swagger/swagger.yaml
