+++ bindings/nodejs/modules/open-intent/lib/rest-server.js
@@ -38,7 +38,7 @@ WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 'use strict';
-var OpenIntentChatbot = require('./lib/chatbot-api/chatbot');
+var OpenIntentChatbot = require('./chatbot-api/chatbot');
 var Q = require('q');
 
 module.exports = function(config, ready) {
