+++ bindings/nodejs/modules/open-intent/project-skeletons/rest-chatbot/README.md
