+++ bindings/nodejs/modules/open-intent/test/chatbot-api/vm-commands.js
@@ -1,0 +1,13 @@
+
+var state = 4;
+
+module.exports = {
+    'command1': function(intentVariables, sessionId, next) {
+        state += 1;
+        next({state: state});
+    },
+    'command2': function(intentVariables, sessionId, next) {
+        state += 2
+        next({state: state});
+    }
+}
