--- bindings/nodejs/modules/open-intent/test/rest-api/helpers/README.md
@@ -1,1 +1,0 @@
-Place your helper tests in this directory.
